Split up the PBO buffer
