+++ common/src/main/java/com/noxcrew/noxesium/config/NoxesiumConfig.java
@@ -17,6 +17,9 @@ public class NoxesiumConfig {
 
     private static final Gson GSON = new GsonBuilder().setPrettyPrinting().create();
 
+    /** Tests whether at least OpenGL 4.4 is present. */
+    public static boolean supportsDynamicUiLimiting = false;
+
     public boolean resetToggleKeys = false;
     public BooleanOrDefault renderMapsInUi = BooleanOrDefault.DEFAULT;
     public boolean showFpsOverlay = false;
@@ -32,6 +35,13 @@ public class NoxesiumConfig {
     public boolean enableDynamicUiLimiting = false;
     public boolean showUiDebugOverlay = false;
     public int maxUiFramerate = 60;
+
+    /**
+     * Whether dynamic UI limiting should be used.
+     */
+    public boolean shouldUseDynamicUiLimiting() {
+        return supportsDynamicUiLimiting && enableUiLimiting;
+    }
 
     /**
      * Returns whether to render maps in the UI.
