+++ common/src/main/resources/assets/noxesium/lang/en_us.json
@@ -36,6 +36,7 @@
   "noxesium.options.max_ui_framerate.tooltip": "Sets the highest framerate at which UI elements are rendered. Only used if `Limit UI Framerate` is ON.",
   "noxesium.options.dynamic_ui_limiting.name": "Dynamic UI Optimization",
   "noxesium.options.dynamic_ui_limiting.tooltip": "§fEnables additional optimizations that dynamically adjust the framerate at which UI elements are rendered, capped at the Max UI Framerate.\n\n§cMay not result in performance improvements for all users, use with caution!",
+  "noxesium.options.dynamic_ui_limiting.disabled": "§cThis setting requires at least OpenGL 4.4 to use.",
   "noxesium.options.ui_debug_overlay.name": "Debug Dynamic UI",
   "noxesium.options.ui_debug_overlay.tooltip": "Adds a small overlay that shows debug information about the dynamic UI rendering optimizations.",
